+++ dataloom/tests/mysql/test_eager_loading_mysql.py
@@ -1117,3 +1117,131 @@ class TestEagerLoadingOnMySQL:
             "supervisor": None,
         }
         conn.close()
+
+    def test_n2n_relations_eager(self):
+        from dataloom import (
+            Loom,
+            Model,
+            Column,
+            PrimaryKeyColumn,
+            TableColumn,
+            ForeignKeyColumn,
+            ColumnValue,
+            Include,
+        )
+        from dataloom.keys import MySQLConfig
+
+        mysql_loom = Loom(
+            dialect="mysql",
+            database=MySQLConfig.database,
+            password=MySQLConfig.password,
+            user=MySQLConfig.user,
+        )
+
+        class Course(Model):
+            __tablename__: TableColumn = TableColumn(name="courses")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+
+        class Student(Model):
+            __tablename__: TableColumn = TableColumn(name="students")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+
+        class StudentCourses(Model):
+            __tablename__: TableColumn = TableColumn(name="students_courses")
+            studentId = ForeignKeyColumn(table=Student, type="int")
+            courseId = ForeignKeyColumn(table=Course, type="int")
+
+        conn, tables = mysql_loom.connect_and_sync(
+            [Student, Course, StudentCourses], force=True
+        )
+
+        # insert the courses
+        mathId = mysql_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="Mathematics")
+        )
+        engId = mysql_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="English")
+        )
+        phyId = mysql_loom.insert_one(
+            instance=Course, values=ColumnValue(name="name", value="Physics")
+        )
+
+        # create students
+
+        stud1 = mysql_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Alice")
+        )
+        stud2 = mysql_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Bob")
+        )
+        stud3 = mysql_loom.insert_one(
+            instance=Student, values=ColumnValue(name="name", value="Lisa")
+        )
+
+        # enrolling students
+        mysql_loom.insert_bulk(
+            instance=StudentCourses,
+            values=[
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=mathId),
+                ],  # enrolling Alice to mathematics
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=phyId),
+                ],  # enrolling Alice to physics
+                [
+                    ColumnValue(name="studentId", value=stud1),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Alice to english
+                [
+                    ColumnValue(name="studentId", value=stud2),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Bob to english
+                [
+                    ColumnValue(name="studentId", value=stud3),
+                    ColumnValue(name="courseId", value=phyId),
+                ],  # enrolling Lisa to physics
+                [
+                    ColumnValue(name="studentId", value=stud3),
+                    ColumnValue(name="courseId", value=engId),
+                ],  # enrolling Lisa to english
+            ],
+        )
+
+        alice = mysql_loom.find_by_pk(
+            Student,
+            pk=stud1,
+            select=["id", "name"],
+            include=Include(model=Course, junction_table=StudentCourses, has="many"),
+        )
+
+        assert alice == {
+            "id": 1,
+            "name": "Alice",
+            "courses": [
+                {"id": 1, "name": "Mathematics"},
+                {"id": 2, "name": "English"},
+                {"id": 3, "name": "Physics"},
+            ],
+        }
+
+        english = mysql_loom.find_by_pk(
+            Course,
+            pk=engId,
+            select=["id", "name"],
+            include=Include(model=Student, junction_table=StudentCourses, has="many"),
+        )
+
+        assert english == {
+            "id": 2,
+            "name": "English",
+            "students": [
+                {"id": 1, "name": "Alice"},
+                {"id": 2, "name": "Bob"},
+                {"id": 3, "name": "Lisa"},
+            ],
+        }
+        conn.close()