+++ dataloom/types/__init__.py
@@ -338,7 +338,7 @@ class Include[Model]:
     ----------
     model : Model
         The model to be included when eger fetching records.
-    junction_table: Model|None, optional
+    junction_table: Model | None, optional
         A junction table is a reference table for models that maps N-N relationship.
     order : list[Order], optional
         The list of order specifications for sorting the included data. Default is an empty list.